+++ src/main/java/com/payflow/exception/DuplicateUpiIdException.java
@@ -1,0 +1,8 @@
+package com.payflow.exception;
+
+public class DuplicateUpiIdException extends PayflowException {
+
+	public DuplicateUpiIdException(String upiId) {
+		super("User already exists with UPI ID: " + upiId);
+	}
+}